[수정] 병합 중 일어난 오류 해결
