--- back/src/main/java/com/golflearn/domain/repository/LessonHistoryRepository.java
@@ -1,37 +1,0 @@
-package com.golflearn.domain.repository;
-
-import java.util.List;
-
-import com.golflearn.dto.LessonLine;
-import com.golflearn.exception.FindException;
-
-public interface LessonHistoryRepository {
-	
-	/**
-	 * 레슨번호에 해당하는 레슨히스토리를 불러온다.
-	 * @param lsnNo
-	 * @return LessonLine
-	 * @throws FindException
-	 */
-	List<LessonLine> selectLessonHistoryByLsnNo(int lsnNo) throws FindException;
-	
-}
-=======
-package com.golflearn.domain.repository;
-
-import java.util.List;
-
-import com.golflearn.dto.LessonLine;
-import com.golflearn.exception.FindException;
-
-public interface LessonHistoryRepository {
-	
-	/**
-	 * 레슨번호에 해당하는 레슨히스토리를 불러온다.
-	 * @param lsnNo
-	 * @return LessonLine
-	 * @throws FindException
-	 */
-	List<LessonLine> selectLessonHistoryByLsnNo(int lsnNo) throws FindException;
-	
-}
