+++ back/src/main/java/com/golflearn/service/LessonService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.golflearn.domain.repository.LessonRepository;
+import com.golflearn.domain.LessonHistoryRepository;
+import com.golflearn.domain.LessonRepository;
 import com.golflearn.dto.Lesson;
-import com.golflearn.domain.repository.LessonHistoryRepository;
 import com.golflearn.dto.LessonLine;
 import com.golflearn.exception.AddException;
