+++ back/src/test/java/com/golflearn/domain/repository/LessonHistoryRepositoryTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-
+import com.golflearn.domain.LessonHistoryRepository;
 import com.golflearn.dto.LessonLine;
 import com.golflearn.exception.FindException;
