+++ back/src/test/java/com/golflearn/domain/repository/UserInfoRepositoryTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import com.golflearn.domain.UserInfoRepository;
 import com.golflearn.dto.UserInfo;
 import com.golflearn.exception.FindException;
 import com.golflearn.exception.ModifyException;
